Create `@babel/helper-environment-visitor`
