+++ packages/babel-helper-create-class-features-plugin/src/misc.ts
@@ -1,7 +1,7 @@
 import { template, traverse, types as t } from "@babel/core";
 import type { File } from "@babel/core";
 import type { NodePath, Scope, Visitor, Binding } from "@babel/traverse";
-import { environmentVisitor } from "@babel/helper-replace-supers";
+import environmentVisitor from "@babel/helper-environment-visitor";
 
 const findBareSupers = traverse.visitors.merge([
   {
