+++ packages/babel-helper-environment-visitor/.npmignore
@@ -1,0 +1,3 @@
+src
+test
+*.log
